Fixed the evolution issue
There was an issue with using a separate file with the PRNG on the device, due to this I were not getting random as was expected. So to solve this I created a unique file with all kernels and now it's working as expected.
TODO: Clean up the code and leave it working like this